refactor: move form validation functions to a class
