+++ client/src/components/registration/RegistrationForm.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import FormError from "../layout/FormError";
-import config from "../../config";
+import FormValidations from "../../services/FormValidations";
 
 const RegistrationForm = () => {
   const [userPayload, setUserPayload] = useState({
@@ -14,53 +14,11 @@ const RegistrationForm = () => {
 
   const [shouldRedirect, setShouldRedirect] = useState(false);
 
-  const validateInput = (payload) => {
-    setErrors({});
-    const { email, username, password, passwordConfirmation } = payload;
-    const emailRegexp = config.validation.email.regexp;
-    let newErrors = {};
-    if (!email.match(emailRegexp)) {
-      newErrors = {
-        ...newErrors,
-        email: "is invalid",
-      };
-    }
-
-    if (username.trim() === "") {
-      newErrors = {
-        ...newErrors,
-        username: "is required",
-      };
-    }
-
-    if (password.trim() === "") {
-      newErrors = {
-        ...newErrors,
-        password: "is required",
-      };
-    }
-
-    if (passwordConfirmation.trim() === "") {
-      newErrors = {
-        ...newErrors,
-        passwordConfirmation: "is required",
-      };
-    } else {
-      if (passwordConfirmation !== password) {
-        newErrors = {
-          ...newErrors,
-          passwordConfirmation: "does not match password",
-        };
-      }
-    }
-
-    setErrors(newErrors)
-    return newErrors
-  };
-
   const onSubmit = async (event) => {
     event.preventDefault();
-    const potentialErrors = validateInput(userPayload);
+    setErrors({})
+    const potentialErrors = FormValidations.registrationForm(userPayload);
+    setErrors(potentialErrors)
     try {
       if (Object.keys(potentialErrors).length === 0) {
         const response = await fetch("/api/v1/users", {
